# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Salon Billing Software API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: staffId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentRequest"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List bills
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: customerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillListResponse"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a new bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get bill details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        membershipType:
          type: string
          nullable: true
        loyaltyPoints:
          type: integer
        totalVisits:
          type: integer
        totalSpend:
          type: number
        lastVisit:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, phone, loyaltyPoints, totalVisits, totalSpend, createdAt]
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            bills:
              type: array
              items:
                $ref: "#/components/schemas/BillSummary"
          required: [bills]
    CustomerListResponse:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [customers, total, page, limit]
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, phone]
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        membershipType:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        staffId:
          type: integer
        staffName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        serviceCategory:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        duration:
          type: integer
        status:
          type: string
          enum: [scheduled, confirmed, in-progress, completed, cancelled]
        notes:
          type: string
          nullable: true
      required: [id, customerId, customerName, customerPhone, staffId, staffName, serviceId, serviceName, serviceCategory, appointmentDate, appointmentTime, duration, status]
    AppointmentListResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required: [appointments]
    CreateAppointmentRequest:
      type: object
      properties:
        customerId:
          type: integer
        staffId:
          type: integer
        serviceId:
          type: integer
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        notes:
          type: string
          nullable: true
      required: [customerId, staffId, serviceId, appointmentDate, appointmentTime]
    UpdateAppointmentRequest:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, confirmed, in-progress, completed, cancelled]
        notes:
          type: string
          nullable: true
    BillItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [service, product]
        itemId:
          type: integer
        name:
          type: string
        staffId:
          type: integer
          nullable: true
        staffName:
          type: string
          nullable: true
        price:
          type: number
        quantity:
          type: integer
        discount:
          type: number
        total:
          type: number
      required: [id, type, itemId, name, price, quantity, discount, total]
    Bill:
      type: object
      properties:
        id:
          type: integer
        billNumber:
          type: string
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/BillItem"
        subtotal:
          type: number
        discountAmount:
          type: number
        taxAmount:
          type: number
        taxPercent:
          type: number
        finalAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, upi, card, wallet]
        couponCode:
          type: string
          nullable: true
        loyaltyPointsUsed:
          type: integer
        status:
          type: string
          enum: [draft, paid, cancelled]
        createdAt:
          type: string
      required: [id, billNumber, items, subtotal, discountAmount, taxAmount, taxPercent, finalAmount, paymentMethod, loyaltyPointsUsed, status, createdAt]
    BillSummary:
      type: object
      properties:
        id:
          type: integer
        billNumber:
          type: string
        finalAmount:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required: [id, billNumber, finalAmount, paymentMethod, status, createdAt]
    BillListResponse:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [bills, total, page, limit]
    CreateBillRequest:
      type: object
      properties:
        customerId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [service, product]
              itemId:
                type: integer
              staffId:
                type: integer
                nullable: true
              quantity:
                type: integer
              discount:
                type: number
            required: [type, itemId, quantity, discount]
        taxPercent:
          type: number
        paymentMethod:
          type: string
          enum: [cash, upi, card, wallet]
        couponCode:
          type: string
          nullable: true
        loyaltyPointsUsed:
          type: integer
        discountAmount:
          type: number
      required: [items, taxPercent, paymentMethod, loyaltyPointsUsed, discountAmount]
...