# lib/api-zod/src/generated/types/appointmentListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon Billing Software API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from "./appointment";

export interface AppointmentListResponse {
  appointments: Appointment[];
}