# lib/api-zod/src/generated/types/billItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon Billing Software API
 * OpenAPI spec version: 0.1.0
 */
import type { BillItemType } from "./billItemType";

export interface BillItem {
  id: number;
  type: BillItemType;
  itemId: number;
  name: string;
  staffId?: number | null;
  staffName?: string | null;
  price: number;
  quantity: number;
  discount: number;
  total: number;
}