# lib/api-zod/src/generated/types/billSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon Billing Software API
 * OpenAPI spec version: 0.1.0
 */

export interface BillSummary {
  id: number;
  billNumber: string;
  finalAmount: number;
  paymentMethod: string;
  status: string;
  createdAt: string;
}